% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kpcoeff_Berez.R
\name{Kpcoeff_Berez}
\alias{Kpcoeff_Berez}
\title{Prediction of tissue-prediction partition coefficient using Berez method}
\usage{
Kpcoeff_Berez(logP, pKa, fup, BP = 1, type = 1, dattype = 0)
}
\arguments{
\item{logP}{Partition coefficient}

\item{pKa}{Acid dissociation constant (for ampholyte or zwitterion, enter Acid pKa first)}

\item{fup}{Plasma unbound fraction}

\item{BP}{Blood-plasma ratio}

\item{type}{Neutral(1)/acid(2)/base(3)/diprotic acid(4)/diprotic base(5)/zwitterion(ampholyte)(6)}

\item{dattype}{Human physiology dataset(0 - original dataset, 1 - unified dataset)}
}
\value{
A list of tissue partition coefficient in each organ
}
\description{
Prediction of tissue-prediction partition coefficient using Berez method
}
\examples{
Kpcoeff_Berez(2.7, 6, 0.9, 1, 1, 0)

}
