% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kpcoeff_Schmitt.R
\name{Kpcoeff_Schmitt}
\alias{Kpcoeff_Schmitt}
\title{Prediction of tissue-prediction partition coefficient using Schmitt method}
\usage{
Kpcoeff_Schmitt(logP, pKa, fup, type = 1, dat)
}
\arguments{
\item{logP}{Partition coefficient}

\item{pKa}{Acid dissociation}

\item{fup}{Plasma unbound fraction}

\item{type}{Neutral/acid/base}

\item{dat}{Human physiology dataset}
}
\value{
A list of tissue partition coefficient in each organ
}
\description{
Prediction of tissue-prediction partition coefficient using Schmitt method
}
\examples{
Kpcoeff_Schmitt(2.7, 6, 0.9, 1, dat_uni)

}
