% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kpcoeff.R
\name{Kpcoeff}
\alias{Kpcoeff}
\title{Calculation of tissue-plasma partition coefficient}
\usage{
Kpcoeff(logP, pKa, fup, BP, type = 3, pred = "P&T", dattype = 0)
}
\arguments{
\item{logP}{partition coefficient (octanol-to-water)}

\item{pKa}{Acid dissociation constant (for ampholyte or zwitterion, enter Acid pKa first)}

\item{fup}{Plasma unbound fraction}

\item{BP}{Blood-to-plasma ratio}

\item{type}{Neutral(1)/acid(2)/base(3)/diprotic acid(4)/diprotic base(5)/zwitterion(ampholyte)(6)}

\item{pred}{Prediction method (Poulin&Theil - "P&T", Rowland&Rowdgers - "R&R", Berezhesky - "Berez", PKsim - "pksim", Schmitt - "Schmitt")}

\item{dattype}{Human physiology dataset(0 - original dataset, 1 - unified dataset)}
}
\value{
A list of partition coefficient in each organ
}
\description{
It calculates tissue-plasma partition coefficient using the specified prediction method and the human physiology dataset
}
\examples{
Kpcoeff(2.6, 2, 0.9, 1.5, 2, "P&T", 0)
Kpcoeff(2.6, c(1.1, 8), 0.5, 1, 6, "R&R", 1)
}
