% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kpcoeff.R
\name{Kpcoeff}
\alias{Kpcoeff}
\title{Calculation of tissue-plasma partition coefficient}
\usage{
Kpcoeff(logP, pKa, fup, BP, type = 3, pred = "P&T", dattype = 0)
}
\arguments{
\item{logP}{partition coefficient (octanol-to-water)}

\item{pKa}{Acid dissociation constant}

\item{fup}{Plasma unbound fraction}

\item{BP}{Blood-to-plasma ratio}

\item{type}{Neutral/base/acid/zwitterion}

\item{pred}{Prediction method}

\item{dattype}{human physiology dataset}
}
\value{
A list of partition coefficient in each organ
}
\description{
It calculates tissue-plasma partition coefficient using the specified prediction method and the human physiology dataset
}
\examples{
Kpcoeff(2.6, 2, 0.9, 1.5, 2, "P&T", 0)

}
