% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kpcoeff_Berez.R
\name{Kpcoeff_Berez}
\alias{Kpcoeff_Berez}
\title{Prediction of tissue-prediction partition coefficient using Berez method}
\usage{
Kpcoeff_Berez(logP, pKa, fup, BP = 1, type = 1, dattype = 0)
}
\arguments{
\item{logP}{Partition coefficient}

\item{pKa}{Acid dissociation}

\item{fup}{Plasma unbound fraction}

\item{BP}{Blood-plasma ratio}

\item{type}{Neutral/acid/base}

\item{dattype}{Human physiology dataset}
}
\value{
A list of tissue partition coefficient in each organ
}
\description{
Prediction of tissue-prediction partition coefficient using Berez method
}
\examples{
Kpcoeff_Berez(2.7, 6, 0.9, 1, 1, 0)

}
