% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kpcoeff_pksim.R
\name{Kpcoeff_pksim}
\alias{Kpcoeff_pksim}
\title{Prediction of tissue-prediction partition coefficient using PK-sim method}
\usage{
Kpcoeff_pksim(logP, fup, dat)
}
\arguments{
\item{logP}{Partition coefficient}

\item{fup}{Plasma unbound fraction}

\item{dat}{Human physiology dataset}
}
\value{
A list of tissue partition coefficient in each organ
}
\description{
Prediction of tissue-prediction partition coefficient using PK-sim method
}
\examples{
Kpcoeff_pksim(2.7,0.9,dat_uni)

}
