% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kpcoeff_RR.R
\name{Kpcoeff_RR}
\alias{Kpcoeff_RR}
\title{Prediction of tissue-prediction partition coefficient using R&R method}
\usage{
Kpcoeff_RR(logP, pKa = 0, fup, BP = 1, type = 1, dat)
}
\arguments{
\item{logP}{Partition coefficient}

\item{pKa}{Acid dissociation}

\item{fup}{Plasma unbound fraction}

\item{BP}{Blood-plasma ratio}

\item{type}{Neutral/acid/base}

\item{dat}{Human physiology dataset}
}
\value{
A list of tissue partition coefficient in each organ
}
\description{
Prediction of tissue-prediction partition coefficient using R&R method
}
\examples{
Kpcoeff_RR(2.7, 6, 0.9, 1, 1, dat_uni)

}
